ai-gateway/promptguard/requirements.txt
fastapi[standard]
uvicorn[standard]
presidio_analyzer[transformers]
presidio_anonymizer
transformers
openai
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc